backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
groq==1.2.0
httpx==0.27.2
sqlalchemy==2.0.30
python-dotenv==1.0.1
pydantic==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
